requirements.txt
aiofiles==23.2.1
aiohttp==3.7.4.post0
aiopokeapi==0.1.8
async-timeout==3.0.1
attrs==21.4.0
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
discord.py==2.1.0
disnake==2.4.0
distlib==0.3.4
filelock==3.4.2
html5lib==1.1
idna==3.3
keyring==21.8.0
lockfile==0.12.2
mpmath==1.2.1
msgpack==1.0.3
multidict==6.0.2
numpy==1.22.1
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.2
platformdirs==2.4.1
poetry==1.1.12
poetry-core==1.0.7
portion==2.3.0
ptyprocess==0.7.0
py-cord==1.7.3
pylev==1.4.0
pyparsing==3.0.7
python-dotenv==0.19.2
PythonTwitchBotFramework==2.3.12
requests==2.27.1
requests-toolbelt==0.9.1
schedule==1.2.1
shellingham==1.4.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
sympy==1.11.1
tomlkit==0.8.0
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.13.0
webencodings==0.5.1
websockets==10.2
yarl==1.7.2
